move invite code to setting and comment drawer
