+++ learn-arai/app/classroom/[slug]/layout.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import SlugContext from '@/components/context/SlugContext';
-import Navbar from '@/components/module/classrooom/navbar/navbar';
-import SubNavBar from '@/components/module/classrooom/sub-navbar/sub-navbar';
+import Navbar from '@/components/module/classroom/navbar/navbar';
+import SubNavBar from '@/components/module/classroom/sub-navbar/sub-navbar';
 import Footer from '@/components/ui/footer/footer';
 
 export default function Layout({
