+++ learn-arai/app/classroom/[slug]/page.tsx
@@ -1,11 +1,9 @@
 'use client';
 
-import CreateInvite from '@/components/module/classrooom/create-invited-code/create-invite-code';
-
 export default function Page() {
     return (
         <>
-            <CreateInvite />
+        // TODO : will insert chat here after merge
         </>
     );
 }
