+++ learn-arai/app/classroom/[slug]/setting/classroom-general.tsx
@@ -4,6 +4,7 @@ import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Separator } from '@/components/ui/separator';
 
 import GroupSection from './group-section';
+import CreateInvite from '@/components/module/classroom/create-invite-code/create-invite-code';
 
 export default function ClassroomGeneral() {
     return (
@@ -17,7 +18,9 @@ export default function ClassroomGeneral() {
             <CardContent className="px-0">
                 <div className="pb-4 px-6">
                     <h3 className="font-medium pb-2 text-lg">Invite Code</h3>
-                    ...
+                    <div className='max-w-xl'>
+                        <CreateInvite />
+                    </div>
                 </div>
 
                 <Separator className="w-[98%] mx-auto" />
