+++ learn-arai/components/module/classroom/delete-group/delete-group.tsx
