+++ learn-arai/components/module/classroom/manage-group-member/manage-group-member.tsx
